backend/requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
PyJWT
email-validator
pandas
joblib
scikit-learn